requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch
torchvision
autocuda
findfile
pyabsa
git+https://github.com/huggingface/diffusers.git
scipy
git+https://github.com/huggingface/transformers.git
ftfy
accelerate
psutil
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
numpy
opencv-python
Pillow
tqdm
realesrgan
